AI lab Type 340 indicator panel.
Photo by Jeff Del Papa.
Source: https://www.reddit.com/r/vintagecomputing/comments/fdpy2s/status_lights_digital_decus_model_30_display_from/

Comments:

The display was a round screen about 16” (40 cm) in diameter,
monochrome green, vector display.

This panel was saved from the one connected to the PDP-6 at the MIT AI
lab/ Project MAC. It was still in service in 1973, when I got a chance
to play spacewar on it. By then they had patched in a small controller
box, you didn’t have to operate the front panel switches, like on the
PDP-1 original.

...

You are correct on the display type, it was a bit of brain fade on my
part. I am not sure when the -6 failed, by the early ‘70 it was
already pretty flakey.

The [Spacewar] controllers I remember were wooden boxes, with a
pressed hardboard front face. they had spin left or right, thrust, a
firing button, and there was a way to get to “hyperspace”, don’t
remember if was it’s own button or the result of hitting a
combination.

The machines were decommissioned in the early ‘80s, and the corpses
(less some parts turned into mementos) went to the Cambridge
electronic surplus dealer “Eli Heffron”. My brother worked there and
snagged the panels for me. I also have a card from one of the
machines, but It is in a box somewhere. They will go to the Computer
Museum when I am gone.

...

The switch panel of the PDP-6 received a similar dated engraving, and
was given to Minsky.

...

I was a regular visitor (a tourist by the local idiom) during my
junior and senior years of high school. I used the ITS machines, but
spent most of my time at the LOGO lab. I continued to use the ITS
machines while I was in college (not MIT), and for the next few years
to get email. For a year or so, I had cstacy as a housemate, which had
good points (he came with a terminal, and 1200 baud modem), but also
some drama, at one weekend we got harassing calls by a then teen aged
Mitnick.

...

I don’t remember the [Dazzle Dart] game, that may have been
later. When I was there, the big development was adding floating point
arithmetic to the language. For the first year I used it, it was
integer only. It was native timesharing, the various terminals all
just gave a language prompt.

I really was into playing with the various bits of hardware they had,
including a 4 voice music maker, some digital breadboards that could
even be connected to the -11 and accessed from the language, and of
course the turtles.

...

As far as I remember, [Cstacy] spent his time keeping the ITS pdp-10
machines running, along with adding to the MIT CADR’s version of the
lisp machine system. In particular he had a role in getting the ITS
machines from the old arpanet 256 fixed host table protocol to actual
TCP.

For the outsiders, a bit of lisp machine history....

By that point, the lisp machine software had branched into 4 separate
streams. Besides the MIT specific version, LMI had one (hardware was
closest to the MIT design), TI had its own hardware , with its own set
of extensions (calling a subroutine was completely different), and
Symbolics, who had the biggest team working on it, their changes
included adding 4 bits to the data paths (36, rather than 32 bits
wide) to make it capable of using full 32 bit addressing. (the others
only did 28 bit addressing) the rest of the bits were for tags.

...

I was playing in the logo lab 73-75 as a high school student. I knew
that SITS was on the pdp-11, I assume on the 11/45 that had run the
standalone multi-user logo, was what turned into the SITS machine, but
I never used it either. By 79, I was working, and didn’t have time to
go hack on logo, I just used the MIT machines as an email host over
dialup.

...

The switches may well have been added by the local denizens. They did
heavily modify things, including adding the memory management hardware
to the KA’s.