NDR plotter code for Muddle.
Attached is a scanned pdf file from a TI silent 700 terminal from 50
years ago, April 16, 1973.  It's a login session from the USC-TIP to
MIT-DMCG.  It contains a listing of Neal Ryans, NDR@DMCG's function
plotting package for imlacs and ards.  This code is from before DEFINE
was used to create functions in Muddle, note the SETG and FUNCTION
being used!  Muddle was like that when I started learning it.  Sorry
about the quality of the scan, but the TI silent 700's thermal output
is 50 plus years old.